Update folder name
