+++ website/src/components/Visualisations/Game/SampleVis.tsx
