+++ website/src/components/Visualisations/IIGO/IIGO.tsx
@@ -7,7 +7,7 @@ const IIGO = () => {
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
       <p className={styles.text}>
-        Self-Organising Multi-Agent Systems 2020
+      <p className={styles.text}>IIGO Visualisation</p>
         </p>
     </div>
   );
